refactor: replace ContactSidePanelController with ContactSidePanel for improved state management
